requirements/base.txt
Django[argon2]==3.2.3
django-environ==0.4.5
django-redis==4.12.1
djangorestframework==3.12.4
drf-spectacular==0.16.0
